Refactor that worked on the first try
