+++ crates/chia-sdk-driver/src/clear_signing.rs
@@ -1,18 +1,24 @@
+mod children;
 mod delegated_spend;
 mod facts;
 mod inner_spend;
-mod linked_spend;
+mod memos;
+mod parsed_spend;
 mod requested_payments;
+mod reveals;
 mod signing;
 mod vault_message;
 mod vault_spend;
 mod vault_transaction;
 
+pub use children::*;
 pub use delegated_spend::*;
 pub use facts::*;
 pub use inner_spend::*;
-pub use linked_spend::*;
+pub use memos::*;
+pub use parsed_spend::*;
 pub use requested_payments::*;
+pub use reveals::*;
 pub use signing::*;
 pub use vault_message::*;
 pub use vault_spend::*;
